Routes

+++ routes/web.php
@@ -16,6 +16,7 @@ Route::get('/', function () {
     return view('welcome');
 })->name('welcome');
 
+
 Route::get('/login', [AuthenticatedSessionController::class, 'create'])->name('login');
 Route::post('/login', [AuthenticatedSessionController::class, 'store']);
 
@@ -26,74 +27,39 @@ Route::get('/dashboard', function () {
 
 require __DIR__.'/auth.php';
 
+Route::middleware('auth')->group(function () {
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+});
+
 //Admin routes
 Route::middleware(['auth', 'admin'])->group(function (){    //'admin' from alias in app.php
     Route::get('/admin/dashboard', [AdminController::class, 'AdminDashboard'])->name('admin.dashboard');
 
-    //route buttons for admin dashboard
+    //route buttons for exclusive for admin 
     Route::get('/employee/register', [EmployeeController::class, 'register'])->name('auth.register');    Route::get('/products', [ProductController::class, 'index'])->name('products.index');
-    Route::get('/products', [ProductController::class, 'index'])->name('product.index');
-    Route::get('/appointments', [AppointmentController::class, 'index'])->name('appointments.index');
-
-    //Routes for Admin profile
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('admin.profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('admin.profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('admin.profile.destroy');
-
     
     
-    // Employee management routes (accessible only by admins)
+    // Employee details management 
     /**get for retrieving, post for storing, put for updating */
-    Route::get('/employee', [EmployeeController::class, 'index'])->name('employee.index');
-    Route::get('/employee/create', [EmployeeController::class, 'create'])->name('employee.create');
-    Route::post('/employee', [EmployeeController::class, 'store'])->name('employee.store');
-    Route::get('/employee/{employee}/edit', [EmployeeController::class, 'edit'])->name('employee.edit');
-    Route::put('/employee/{employee}/update', [EmployeeController::class, 'update'])->name('employee.update');
-    Route::delete('/employee/{employee}/destroy', [EmployeeController::class, 'destroy'])->name('employee.destroy');
-
-    //product routes for admin
-    Route::get ('/product',[ProductController::class,'index'])->name('product.index');/*path where the view file is located*/
-    Route::get ('/product/create',[ProductController::class,'create'])->name('product.create');
-    Route::post ('/product',[ProductController::class,'store'])->name('product.store');
-    Route::get ('/product/{product}/edit',[ProductController::class,'edit'])->name('product.edit'); //edit is the fuction
-    Route::put ('/product/{product}/update',[ProductController::class,'update'])->name('product.update'); 
-    Route::delete ('/product/{product}/delete',[ProductController::class,'delete'])->name('product.delete'); 
-
+    Route::get('/admin/employee', [EmployeeController::class, 'index'])->name('admin.employee.index');
+    Route::get('/admin/employee/create', [EmployeeController::class, 'create'])->name('admin.employee.create');
+    Route::post('/admin/employee', [EmployeeController::class, 'store'])->name('admin.employee.store');
+    Route::get('/admin/employee/{employee}/edit', [EmployeeController::class, 'edit'])->name('admin.employee.edit');
+    Route::put('/admin/employee/{employee}/update', [EmployeeController::class, 'update'])->name('admin.employee.update');
+    Route::delete('/admin/employee/{employee}/destroy', [EmployeeController::class, 'destroy'])->name('admin.employee.destroy');
 });
 
 //Employee routes
 Route::middleware(['auth', 'employee'])->group(function () {    //'employee' from alias in app.php
     Route::get('/employee/dashboard', [EmployeeController::class, 'EmployeeDashboard'])->name('employee.dashboard');
 
-    //employee dashboard routes
-    Route::get('/appointments', [AppointmentController::class, 'index'])->name('appointments.index');
-    Route::get('/products', [ProductController::class, 'index'])->name('product.index');
-
-    //profile routes for employee
-    Route::get('/employee/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/employee/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/employee/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-    
 });
 
-//Customer routes
-Route::middleware(['auth', 'customer'])->group(function () {    //'customer' from alias in app.php
-    Route::get('/customer', [CustomerController::class, 'customerdashboard'])->name('customer.dashboard');
+// admin and employee  middleware
+Route::middleware(['admin_or_employee'])->group(function () {
 
-    // Profile routes for customers
-    //customer/login
-    //need a Home for header
-    Route::get('/customer/profile', [CustomerController::class, 'profile'])->name('customer.profile');
-    Route::post('/customer/update-profile', [CustomerController::class, 'updateProfile'])->name('customer.updateProfile');
-    Route::delete('/customer/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-
-    //appointment routes
-    Route::get('appointments/create', [AppointmentController::class, 'create'])->name('appointments.create');
-    Route::post('appointments', [AppointmentController::class, 'store'])->name('appointments.store');
-});
-
-// admin and customer middleware
-Route::middleware(['auth', 'admin_or_employee'])->group(function () {
     //product routes
     Route::get ('/product',[ProductController::class,'index'])->name('product.index');/*path where the view file is located*/
     Route::get ('/product/create',[ProductController::class,'create'])->name('product.create');
@@ -110,8 +76,26 @@ Route::middleware(['auth', 'admin_or_employee'])->group(function () {
 
 });
 
+//Customer routes
+Route::middleware(['auth', 'customer'])->group(function () {    //'customer' from alias in app.php
+    Route::get('/customer', [CustomerController::class, 'customerdashboard'])->name('customer.dashboard');
+
+    // Profile routes for customers
+    //need a Home for header
+    Route::get('/customer/profile', [CustomerController::class, 'profile'])->name('customer.profile');
+    Route::post('/customer/update-profile', [CustomerController::class, 'updateProfile'])->name('customer.updateProfile');
+    Route::delete('/customer/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+
+    //appointment routes
+    Route::get('appointments/create', [AppointmentController::class, 'create'])->name('appointments.create');
+    Route::post('appointments', [AppointmentController::class, 'store'])->name('appointments.store');
+});
+
+
+
 // Publicly accessible routes
 Route::get('/homepage', [CustomerController::class, 'index'])->name('customer.index');
+
      
-
-
+ 
+
